add description_vacancy.js

+++ src/main/java/com/wanted/needswork/DTO/request/ResponseDTO.java
@@ -3,13 +3,14 @@ package com.wanted.needswork.DTO.request;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigInteger;
 
 
 @AllArgsConstructor
 public class ResponseDTO {
 
     @Getter
-    private Integer vacancy_id;
+    private BigInteger vacancy_id;
     @Getter
     private Integer job_seeker_id;
